stat/uucore: refactor - move fsext.rs to uucore
