+++ src/uucore/Cargo.toml
@@ -38,6 +38,7 @@ default = []
 encoding = ["data-encoding", "thiserror"]
 entries = ["libc"]
 fs = ["libc"]
+fsext = ["libc", "time"]
 mode = ["libc"]
 parse_time = []
 perms = ["libc"]
